+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog Format
 NOTE: all lowercase and no ending punctuation marks
 -->
 
+- [CY] - [#7](https://github.com/cyril-deguzman/studyuwu/issues/7) app(service): handle invalid creation of bookauthor requests
 - [CY] - [#5](https://github.com/cyril-deguzman/studyuwu/issues/5) app(all): refactor whole app
 - [CY] - [#4](https://github.com/cyril-deguzman/studyuwu/issues/4) app(service): add or remove author to and from books
 - [CY] - [#3](https://github.com/cyril-deguzman/studyuwu/issues/3) app(module): implement basic crud operations, validation, and exception filters for books module